+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 
 ## v0.13-SNAPSHOT
 - added a load button.
-
+- added a route '/api/load'.
+- added a client function, client/load-messages
 
 ## v0.12.116 / 2024-06-28
 - only admin can use Ctrl+U.