+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - added a load button.
 - added a route '/api/load'.
 - added a client function, client/load-messages
+- added a incomplete function, client/replace-content
+
 
 ## v0.12.116 / 2024-06-28
 - only admin can use Ctrl+U.